+++ internal/dbaas/cluster.go
@@ -70,22 +70,6 @@ func (s *ClusterTool) getCluster(ctx context.Context, req mcp.CallToolRequest) (
 	return mcp.NewToolResultText(string(jsonCluster)), nil
 }
 
-func (s *ClusterTool) getCA(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	id, ok := req.GetArguments()["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	ca, _, err := s.client.Databases.GetCA(ctx, id)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonCA, err := json.MarshalIndent(ca, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonCA)), nil
-}
-
 func (s *ClusterTool) createCluster(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
 	args := req.GetArguments()
 
@@ -198,6 +182,22 @@ func (s *ClusterTool) migrateCluster(ctx context.Context, req mcp.CallToolReques
 	return mcp.NewToolResultText("Cluster migration initiated successfully"), nil
 }
 
+func (s *ClusterTool) getCA(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	id, ok := req.GetArguments()["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+	ca, _, err := s.client.Databases.GetCA(ctx, id)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	jsonCA, err := json.MarshalIndent(ca, "", "  ")
+	if err != nil {
+		return nil, fmt.Errorf("marshal error: %w", err)
+	}
+	return mcp.NewToolResultText(string(jsonCA)), nil
+}
+
 func (s *ClusterTool) updateMaintenance(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
 	args := req.GetArguments()
 	id, ok := args["ID"].(string)
@@ -271,152 +271,6 @@ func (s *ClusterTool) listBackups(ctx context.Context, req mcp.CallToolRequest) 
 	return mcp.NewToolResultText(string(jsonBackups)), nil
 }
 
-func (s *ClusterTool) getUser(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	user, ok := args["user"].(string)
-	if !ok || user == "" {
-		return mcp.NewToolResultError("User name is required"), nil
-	}
-
-	dbUser, _, err := s.client.Databases.GetUser(ctx, id, user)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonUser, err := json.MarshalIndent(dbUser, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonUser)), nil
-}
-
-func (s *ClusterTool) listUsers(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-
-	// Optional pagination
-	page := 0
-	if pStr, ok := args["page"].(string); ok && pStr != "" {
-		if p, err := strconv.Atoi(pStr); err == nil {
-			page = p
-		}
-	}
-	perPage := 0
-	if ppStr, ok := args["per_page"].(string); ok && ppStr != "" {
-		if pp, err := strconv.Atoi(ppStr); err == nil {
-			perPage = pp
-		}
-	}
-	var opts *godo.ListOptions
-	if page > 0 || perPage > 0 {
-		opts = &godo.ListOptions{Page: page, PerPage: perPage}
-	}
-
-	users, _, err := s.client.Databases.ListUsers(ctx, id, opts)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonUsers, err := json.MarshalIndent(users, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonUsers)), nil
-}
-
-func (s *ClusterTool) createUser(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("User name is required"), nil
-	}
-
-	createReq := &godo.DatabaseCreateUserRequest{Name: name}
-
-	if plugin, ok := args["mysql_auth_plugin"].(string); ok && plugin != "" {
-		createReq.MySQLSettings = &godo.DatabaseMySQLUserSettings{AuthPlugin: plugin}
-	}
-
-	if settingsStr, ok := args["settings_json"].(string); ok && settingsStr != "" {
-		var settings godo.DatabaseUserSettings
-		err := json.Unmarshal([]byte(settingsStr), &settings)
-		if err != nil {
-			return mcp.NewToolResultError("Invalid settings_json: " + err.Error()), nil
-		}
-		createReq.Settings = &settings
-	}
-
-	dbUser, _, err := s.client.Databases.CreateUser(ctx, id, createReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonUser, err := json.MarshalIndent(dbUser, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonUser)), nil
-}
-
-func (s *ClusterTool) updateUser(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	user, ok := args["user"].(string)
-	if !ok || user == "" {
-		return mcp.NewToolResultError("User name is required"), nil
-	}
-
-	updateReq := &godo.DatabaseUpdateUserRequest{}
-
-	if settingsStr, ok := args["settings_json"].(string); ok && settingsStr != "" {
-		var settings godo.DatabaseUserSettings
-		err := json.Unmarshal([]byte(settingsStr), &settings)
-		if err != nil {
-			return mcp.NewToolResultError("Invalid settings_json: " + err.Error()), nil
-		}
-		updateReq.Settings = &settings
-	}
-
-	dbUser, _, err := s.client.Databases.UpdateUser(ctx, id, user, updateReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonUser, err := json.MarshalIndent(dbUser, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonUser)), nil
-}
-
-func (s *ClusterTool) deleteUser(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	user, ok := args["user"].(string)
-	if !ok || user == "" {
-		return mcp.NewToolResultError("User name is required"), nil
-	}
-
-	_, err := s.client.Databases.DeleteUser(ctx, id, user)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("User deleted successfully"), nil
-}
-
 func (s *ClusterTool) resetUserAuth(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
 	args := req.GetArguments()
 	id, ok := args["ID"].(string)
@@ -453,405 +307,6 @@ func (s *ClusterTool) resetUserAuth(ctx context.Context, req mcp.CallToolRequest
 	return mcp.NewToolResultText(string(jsonUser)), nil
 }
 
-func (s *ClusterTool) listDBs(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-
-	// Optional pagination
-	page := 0
-	if pStr, ok := args["page"].(string); ok && pStr != "" {
-		if p, err := strconv.Atoi(pStr); err == nil {
-			page = p
-		}
-	}
-	perPage := 0
-	if ppStr, ok := args["per_page"].(string); ok && ppStr != "" {
-		if pp, err := strconv.Atoi(ppStr); err == nil {
-			perPage = pp
-		}
-	}
-	var opts *godo.ListOptions
-	if page > 0 || perPage > 0 {
-		opts = &godo.ListOptions{Page: page, PerPage: perPage}
-	}
-
-	dbs, _, err := s.client.Databases.ListDBs(ctx, id, opts)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonDBs, err := json.MarshalIndent(dbs, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonDBs)), nil
-}
-
-func (s *ClusterTool) createDB(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Database name is required"), nil
-	}
-
-	createReq := &godo.DatabaseCreateDBRequest{Name: name}
-	db, _, err := s.client.Databases.CreateDB(ctx, id, createReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonDB, err := json.MarshalIndent(db, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonDB)), nil
-}
-
-func (s *ClusterTool) getDB(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Database name is required"), nil
-	}
-	db, _, err := s.client.Databases.GetDB(ctx, id, name)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonDB, err := json.MarshalIndent(db, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonDB)), nil
-}
-
-func (s *ClusterTool) deleteDB(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Database name is required"), nil
-	}
-	_, err := s.client.Databases.DeleteDB(ctx, id, name)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Database deleted successfully"), nil
-}
-
-func (s *ClusterTool) listPools(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-
-	// Optional pagination
-	page := 0
-	if pStr, ok := args["page"].(string); ok && pStr != "" {
-		if p, err := strconv.Atoi(pStr); err == nil {
-			page = p
-		}
-	}
-	perPage := 0
-	if ppStr, ok := args["per_page"].(string); ok && ppStr != "" {
-		if pp, err := strconv.Atoi(ppStr); err == nil {
-			perPage = pp
-		}
-	}
-	var opts *godo.ListOptions
-	if page > 0 || perPage > 0 {
-		opts = &godo.ListOptions{Page: page, PerPage: perPage}
-	}
-
-	pools, _, err := s.client.Databases.ListPools(ctx, id, opts)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonPools, err := json.MarshalIndent(pools, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonPools)), nil
-}
-
-func (s *ClusterTool) createPool(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	user, ok := args["user"].(string)
-	if !ok || user == "" {
-		return mcp.NewToolResultError("User is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Pool name is required"), nil
-	}
-	database, ok := args["database"].(string)
-	if !ok || database == "" {
-		return mcp.NewToolResultError("Database is required"), nil
-	}
-	mode, ok := args["mode"].(string)
-	if !ok || mode == "" {
-		return mcp.NewToolResultError("Mode is required"), nil
-	}
-	sizeF, ok := args["size"].(float64)
-	if !ok {
-		return mcp.NewToolResultError("Size is required and must be a number"), nil
-	}
-	size := int(sizeF)
-
-	createReq := &godo.DatabaseCreatePoolRequest{
-		User:     user,
-		Name:     name,
-		Database: database,
-		Mode:     mode,
-		Size:     size,
-	}
-	pool, _, err := s.client.Databases.CreatePool(ctx, id, createReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonPool, err := json.MarshalIndent(pool, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonPool)), nil
-}
-
-func (s *ClusterTool) getPool(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Pool name is required"), nil
-	}
-	pool, _, err := s.client.Databases.GetPool(ctx, id, name)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonPool, err := json.MarshalIndent(pool, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonPool)), nil
-}
-
-func (s *ClusterTool) deletePool(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Pool name is required"), nil
-	}
-	_, err := s.client.Databases.DeletePool(ctx, id, name)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Pool deleted successfully"), nil
-}
-
-func (s *ClusterTool) updatePool(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Pool name is required"), nil
-	}
-	database, ok := args["database"].(string)
-	if !ok || database == "" {
-		return mcp.NewToolResultError("Database is required"), nil
-	}
-	mode, ok := args["mode"].(string)
-	if !ok || mode == "" {
-		return mcp.NewToolResultError("Mode is required"), nil
-	}
-	sizeF, ok := args["size"].(float64)
-	if !ok {
-		return mcp.NewToolResultError("Size is required and must be a number"), nil
-	}
-	size := int(sizeF)
-	user, _ := args["user"].(string)
-
-	updateReq := &godo.DatabaseUpdatePoolRequest{
-		User:     user,
-		Database: database,
-		Mode:     mode,
-		Size:     size,
-	}
-	_, err := s.client.Databases.UpdatePool(ctx, id, name, updateReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Pool updated successfully"), nil
-}
-
-func (s *ClusterTool) getReplica(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Replica name is required"), nil
-	}
-	replica, _, err := s.client.Databases.GetReplica(ctx, id, name)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonReplica, err := json.MarshalIndent(replica, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonReplica)), nil
-}
-
-func (s *ClusterTool) listReplicas(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-
-	// Optional pagination
-	page := 0
-	if pStr, ok := args["page"].(string); ok && pStr != "" {
-		if p, err := strconv.Atoi(pStr); err == nil {
-			page = p
-		}
-	}
-	perPage := 0
-	if ppStr, ok := args["per_page"].(string); ok && ppStr != "" {
-		if pp, err := strconv.Atoi(ppStr); err == nil {
-			perPage = pp
-		}
-	}
-	var opts *godo.ListOptions
-	if page > 0 || perPage > 0 {
-		opts = &godo.ListOptions{Page: page, PerPage: perPage}
-	}
-
-	replicas, _, err := s.client.Databases.ListReplicas(ctx, id, opts)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonReplicas, err := json.MarshalIndent(replicas, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonReplicas)), nil
-}
-
-func (s *ClusterTool) createReplica(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Replica name is required"), nil
-	}
-	region, ok := args["region"].(string)
-	if !ok || region == "" {
-		return mcp.NewToolResultError("Replica region is required"), nil
-	}
-	size, ok := args["size"].(string)
-	if !ok || size == "" {
-		return mcp.NewToolResultError("Replica size is required"), nil
-	}
-	privateNetworkUUID, _ := args["private_network_uuid"].(string)
-	tags := []string{}
-	if tagsRaw, ok := args["tags"].(string); ok && tagsRaw != "" {
-		for _, t := range strings.Split(tagsRaw, ",") {
-			t = strings.TrimSpace(t)
-			if t != "" {
-				tags = append(tags, t)
-			}
-		}
-	}
-	storageSizeMib := uint64(0)
-	if ssm, ok := args["storage_size_mib"].(float64); ok {
-		storageSizeMib = uint64(ssm)
-	}
-
-	createReq := &godo.DatabaseCreateReplicaRequest{
-		Name:               name,
-		Region:             region,
-		Size:               size,
-		PrivateNetworkUUID: privateNetworkUUID,
-		Tags:               tags,
-		StorageSizeMib:     storageSizeMib,
-	}
-
-	replica, _, err := s.client.Databases.CreateReplica(ctx, id, createReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonReplica, err := json.MarshalIndent(replica, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonReplica)), nil
-}
-
-func (s *ClusterTool) deleteReplica(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Replica name is required"), nil
-	}
-	_, err := s.client.Databases.DeleteReplica(ctx, id, name)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Replica deleted successfully"), nil
-}
-
-func (s *ClusterTool) promoteReplicaToPrimary(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Replica name is required"), nil
-	}
-	_, err := s.client.Databases.PromoteReplicaToPrimary(ctx, id, name)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Replica promoted to primary successfully"), nil
-}
-
 func (s *ClusterTool) getEvictionPolicy(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
 	args := req.GetArguments()
 	id, ok := args["ID"].(string)
@@ -882,317 +337,6 @@ func (s *ClusterTool) setEvictionPolicy(ctx context.Context, req mcp.CallToolReq
 	return mcp.NewToolResultText("Eviction policy set successfully"), nil
 }
 
-func (s *ClusterTool) getSQLMode(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	mode, _, err := s.client.Databases.GetSQLMode(ctx, id)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText(mode), nil
-}
-
-func (s *ClusterTool) setSQLMode(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	modesStr, ok := args["modes"].(string)
-	if !ok || modesStr == "" {
-		return mcp.NewToolResultError("SQL modes are required (comma-separated)"), nil
-	}
-	modes := []string{}
-	for _, m := range strings.Split(modesStr, ",") {
-		m = strings.TrimSpace(m)
-		if m != "" {
-			modes = append(modes, m)
-		}
-	}
-	_, err := s.client.Databases.SetSQLMode(ctx, id, modes...)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("SQL mode set successfully"), nil
-}
-
-func (s *ClusterTool) getFirewallRules(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	rules, _, err := s.client.Databases.GetFirewallRules(ctx, id)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonRules, err := json.MarshalIndent(rules, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonRules)), nil
-}
-
-func (s *ClusterTool) updateFirewallRules(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	rulesStr, ok := args["rules_json"].(string)
-	if !ok || rulesStr == "" {
-		return mcp.NewToolResultError("rules_json is required (JSON array of firewall rules)"), nil
-	}
-	var rules []*godo.DatabaseFirewallRule
-	err := json.Unmarshal([]byte(rulesStr), &rules)
-	if err != nil {
-		return mcp.NewToolResultError("Invalid rules_json: " + err.Error()), nil
-	}
-	updateReq := &godo.DatabaseUpdateFirewallRulesRequest{Rules: rules}
-	_, err = s.client.Databases.UpdateFirewallRules(ctx, id, updateReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Firewall rules updated successfully"), nil
-}
-
-func (s *ClusterTool) getPostgreSQLConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	cfg, _, err := s.client.Databases.GetPostgreSQLConfig(ctx, id)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonCfg, err := json.MarshalIndent(cfg, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonCfg)), nil
-}
-
-func (s *ClusterTool) updatePostgreSQLConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	configStr, ok := args["config_json"].(string)
-	if !ok || configStr == "" {
-		return mcp.NewToolResultError("config_json is required (JSON for PostgreSQLConfig)"), nil
-	}
-	var config godo.PostgreSQLConfig
-	err := json.Unmarshal([]byte(configStr), &config)
-	if err != nil {
-		return mcp.NewToolResultError("Invalid config_json: " + err.Error()), nil
-	}
-	_, err = s.client.Databases.UpdatePostgreSQLConfig(ctx, id, &config)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("PostgreSQL config updated successfully"), nil
-}
-
-func (s *ClusterTool) getRedisConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	cfg, _, err := s.client.Databases.GetRedisConfig(ctx, id)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonCfg, err := json.MarshalIndent(cfg, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonCfg)), nil
-}
-
-func (s *ClusterTool) updateRedisConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	configStr, ok := args["config_json"].(string)
-	if !ok || configStr == "" {
-		return mcp.NewToolResultError("config_json is required (JSON for RedisConfig)"), nil
-	}
-	var config godo.RedisConfig
-	err := json.Unmarshal([]byte(configStr), &config)
-	if err != nil {
-		return mcp.NewToolResultError("Invalid config_json: " + err.Error()), nil
-	}
-	_, err = s.client.Databases.UpdateRedisConfig(ctx, id, &config)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Redis config updated successfully"), nil
-}
-
-func (s *ClusterTool) getMySQLConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	cfg, _, err := s.client.Databases.GetMySQLConfig(ctx, id)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonCfg, err := json.MarshalIndent(cfg, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonCfg)), nil
-}
-
-func (s *ClusterTool) updateMySQLConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	configStr, ok := args["config_json"].(string)
-	if !ok || configStr == "" {
-		return mcp.NewToolResultError("config_json is required (JSON for MySQLConfig)"), nil
-	}
-	var config godo.MySQLConfig
-	err := json.Unmarshal([]byte(configStr), &config)
-	if err != nil {
-		return mcp.NewToolResultError("Invalid config_json: " + err.Error()), nil
-	}
-	_, err = s.client.Databases.UpdateMySQLConfig(ctx, id, &config)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("MySQL config updated successfully"), nil
-}
-
-func (s *ClusterTool) getMongoDBConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	cfg, _, err := s.client.Databases.GetMongoDBConfig(ctx, id)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonCfg, err := json.MarshalIndent(cfg, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonCfg)), nil
-}
-
-func (s *ClusterTool) updateMongoDBConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	configStr, ok := args["config_json"].(string)
-	if !ok || configStr == "" {
-		return mcp.NewToolResultError("config_json is required (JSON for MongoDBConfig)"), nil
-	}
-	var config godo.MongoDBConfig
-	err := json.Unmarshal([]byte(configStr), &config)
-	if err != nil {
-		return mcp.NewToolResultError("Invalid config_json: " + err.Error()), nil
-	}
-	_, err = s.client.Databases.UpdateMongoDBConfig(ctx, id, &config)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("MongoDB config updated successfully"), nil
-}
-
-func (s *ClusterTool) getOpensearchConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	cfg, _, err := s.client.Databases.GetOpensearchConfig(ctx, id)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonCfg, err := json.MarshalIndent(cfg, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonCfg)), nil
-}
-
-func (s *ClusterTool) updateOpensearchConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	configStr, ok := args["config_json"].(string)
-	if !ok || configStr == "" {
-		return mcp.NewToolResultError("config_json is required (JSON for OpensearchConfig)"), nil
-	}
-	var config godo.OpensearchConfig
-	err := json.Unmarshal([]byte(configStr), &config)
-	if err != nil {
-		return mcp.NewToolResultError("Invalid config_json: " + err.Error()), nil
-	}
-	_, err = s.client.Databases.UpdateOpensearchConfig(ctx, id, &config)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Opensearch config updated successfully"), nil
-}
-
-func (s *ClusterTool) getKafkaConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	cfg, _, err := s.client.Databases.GetKafkaConfig(ctx, id)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonCfg, err := json.MarshalIndent(cfg, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonCfg)), nil
-}
-
-func (s *ClusterTool) updateKafkaConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	configStr, ok := args["config_json"].(string)
-	if !ok || configStr == "" {
-		return mcp.NewToolResultError("config_json is required (JSON for KafkaConfig)"), nil
-	}
-	var config godo.KafkaConfig
-	err := json.Unmarshal([]byte(configStr), &config)
-	if err != nil {
-		return mcp.NewToolResultError("Invalid config_json: " + err.Error()), nil
-	}
-	_, err = s.client.Databases.UpdateKafkaConfig(ctx, id, &config)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Kafka config updated successfully"), nil
-}
-
 func (s *ClusterTool) listOptions(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
 	options, _, err := s.client.Databases.ListOptions(ctx)
 	if err != nil {
@@ -1221,231 +365,6 @@ func (s *ClusterTool) upgradeMajorVersion(ctx context.Context, req mcp.CallToolR
 		return mcp.NewToolResultErrorFromErr("api error", err), nil
 	}
 	return mcp.NewToolResultText("Major version upgrade initiated successfully"), nil
-}
-
-func (s *ClusterTool) listTopics(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-
-	opts := &godo.ListOptions{}
-	if pStr, ok := args["page"].(string); ok && pStr != "" {
-		if p, err := strconv.Atoi(pStr); err == nil {
-			opts.Page = p
-		}
-	}
-	if ppStr, ok := args["per_page"].(string); ok && ppStr != "" {
-		if pp, err := strconv.Atoi(ppStr); err == nil {
-			opts.PerPage = pp
-		}
-	}
-	if wpStr, ok := args["with_projects"].(string); ok && wpStr != "" {
-		if wp, err := strconv.ParseBool(wpStr); err == nil {
-			opts.WithProjects = wp
-		}
-	}
-	if odStr, ok := args["only_deployed"].(string); ok && odStr != "" {
-		if od, err := strconv.ParseBool(odStr); err == nil {
-			opts.Deployed = od
-		}
-	}
-	if poStr, ok := args["public_only"].(string); ok && poStr != "" {
-		if po, err := strconv.ParseBool(poStr); err == nil {
-			opts.PublicOnly = po
-		}
-	}
-	if ucStr, ok := args["usecases"].(string); ok && ucStr != "" {
-		ucList := []string{}
-		for _, u := range strings.Split(ucStr, ",") {
-			u = strings.TrimSpace(u)
-			if u != "" {
-				ucList = append(ucList, u)
-			}
-		}
-		opts.Usecases = ucList
-	}
-
-	topics, _, err := s.client.Databases.ListTopics(ctx, id, opts)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonTopics, err := json.MarshalIndent(topics, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonTopics)), nil
-}
-
-func (s *ClusterTool) createTopic(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Topic name is required"), nil
-	}
-
-	var partitionCount *uint32
-	if pcStr, ok := args["partition_count"].(string); ok && pcStr != "" {
-		if pc, err := strconv.ParseUint(pcStr, 10, 32); err == nil {
-			pc32 := uint32(pc)
-			partitionCount = &pc32
-		}
-	}
-	var replicationFactor *uint32
-	if rfStr, ok := args["replication_factor"].(string); ok && rfStr != "" {
-		if rf, err := strconv.ParseUint(rfStr, 10, 32); err == nil {
-			rf32 := uint32(rf)
-			replicationFactor = &rf32
-		}
-	}
-
-	var topicConfig *godo.TopicConfig
-	if cfgStr, ok := args["config_json"].(string); ok && cfgStr != "" {
-		var cfg godo.TopicConfig
-		err := json.Unmarshal([]byte(cfgStr), &cfg)
-		if err != nil {
-			return mcp.NewToolResultError("Invalid config_json: " + err.Error()), nil
-		}
-		topicConfig = &cfg
-	}
-
-	createReq := &godo.DatabaseCreateTopicRequest{
-		Name:              name,
-		PartitionCount:    partitionCount,
-		ReplicationFactor: replicationFactor,
-		Config:            topicConfig,
-	}
-	topic, _, err := s.client.Databases.CreateTopic(ctx, id, createReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonTopic, err := json.MarshalIndent(topic, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonTopic)), nil
-}
-
-func (s *ClusterTool) getTopic(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Topic name is required"), nil
-	}
-	topic, _, err := s.client.Databases.GetTopic(ctx, id, name)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonTopic, err := json.MarshalIndent(topic, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonTopic)), nil
-}
-
-func (s *ClusterTool) deleteTopic(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Topic name is required"), nil
-	}
-	_, err := s.client.Databases.DeleteTopic(ctx, id, name)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Topic deleted successfully"), nil
-}
-
-func (s *ClusterTool) updateTopic(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Topic name is required"), nil
-	}
-
-	var partitionCount *uint32
-	if pcStr, ok := args["partition_count"].(string); ok && pcStr != "" {
-		if pc, err := strconv.ParseUint(pcStr, 10, 32); err == nil {
-			pc32 := uint32(pc)
-			partitionCount = &pc32
-		}
-	}
-	var replicationFactor *uint32
-	if rfStr, ok := args["replication_factor"].(string); ok && rfStr != "" {
-		if rf, err := strconv.ParseUint(rfStr, 10, 32); err == nil {
-			rf32 := uint32(rf)
-			replicationFactor = &rf32
-		}
-	}
-
-	var topicConfig *godo.TopicConfig
-	if cfgStr, ok := args["config_json"].(string); ok && cfgStr != "" {
-		var cfg godo.TopicConfig
-		err := json.Unmarshal([]byte(cfgStr), &cfg)
-		if err != nil {
-			return mcp.NewToolResultError("Invalid config_json: " + err.Error()), nil
-		}
-		topicConfig = &cfg
-	}
-
-	updateReq := &godo.DatabaseUpdateTopicRequest{
-		PartitionCount:    partitionCount,
-		ReplicationFactor: replicationFactor,
-		Config:            topicConfig,
-	}
-	_, err := s.client.Databases.UpdateTopic(ctx, id, name, updateReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Topic updated successfully"), nil
-}
-
-func (s *ClusterTool) getMetricsCredentials(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	creds, _, err := s.client.Databases.GetMetricsCredentials(ctx)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonCreds, err := json.MarshalIndent(creds, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonCreds)), nil
-}
-
-func (s *ClusterTool) updateMetricsCredentials(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	credsStr, ok := args["credentials_json"].(string)
-	if !ok || credsStr == "" {
-		return mcp.NewToolResultError("credentials_json is required (JSON for DatabaseMetricsCredentials)"), nil
-	}
-	var creds godo.DatabaseMetricsCredentials
-	err := json.Unmarshal([]byte(credsStr), &creds)
-	if err != nil {
-		return mcp.NewToolResultError("Invalid credentials_json: " + err.Error()), nil
-	}
-	updateReq := &godo.DatabaseUpdateMetricsCredentialsRequest{Credentials: &creds}
-	_, err = s.client.Databases.UpdateMetricsCredentials(ctx, updateReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Metrics credentials updated successfully"), nil
 }
 
 func (s *ClusterTool) listDatabaseEvents(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
@@ -1501,296 +420,6 @@ func (s *ClusterTool) listDatabaseEvents(ctx context.Context, req mcp.CallToolRe
 		return nil, fmt.Errorf("marshal error: %w", err)
 	}
 	return mcp.NewToolResultText(string(jsonEvents)), nil
-}
-
-func (s *ClusterTool) listIndexes(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-
-	opts := &godo.ListOptions{}
-	if pStr, ok := args["page"].(string); ok && pStr != "" {
-		if p, err := strconv.Atoi(pStr); err == nil {
-			opts.Page = p
-		}
-	}
-	if ppStr, ok := args["per_page"].(string); ok && ppStr != "" {
-		if pp, err := strconv.Atoi(ppStr); err == nil {
-			opts.PerPage = pp
-		}
-	}
-	if wpStr, ok := args["with_projects"].(string); ok && wpStr != "" {
-		if wp, err := strconv.ParseBool(wpStr); err == nil {
-			opts.WithProjects = wp
-		}
-	}
-	if odStr, ok := args["only_deployed"].(string); ok && odStr != "" {
-		if od, err := strconv.ParseBool(odStr); err == nil {
-			opts.Deployed = od
-		}
-	}
-	if poStr, ok := args["public_only"].(string); ok && poStr != "" {
-		if po, err := strconv.ParseBool(poStr); err == nil {
-			opts.PublicOnly = po
-		}
-	}
-	if ucStr, ok := args["usecases"].(string); ok && ucStr != "" {
-		ucList := []string{}
-		for _, u := range strings.Split(ucStr, ",") {
-			u = strings.TrimSpace(u)
-			if u != "" {
-				ucList = append(ucList, u)
-			}
-		}
-		opts.Usecases = ucList
-	}
-
-	indexes, _, err := s.client.Databases.ListIndexes(ctx, id, opts)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonIndexes, err := json.MarshalIndent(indexes, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonIndexes)), nil
-}
-
-func (s *ClusterTool) deleteIndex(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("Index name is required"), nil
-	}
-	_, err := s.client.Databases.DeleteIndex(ctx, id, name)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Index deleted successfully"), nil
-}
-
-func (s *ClusterTool) createLogsink(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	name, ok := args["sink_name"].(string)
-	if !ok || name == "" {
-		return mcp.NewToolResultError("sink_name is required"), nil
-	}
-	typeStr, ok := args["sink_type"].(string)
-	if !ok || typeStr == "" {
-		return mcp.NewToolResultError("sink_type is required"), nil
-	}
-	configStr, ok := args["config_json"].(string)
-	if !ok || configStr == "" {
-		return mcp.NewToolResultError("config_json is required (JSON for DatabaseLogsinkConfig)"), nil
-	}
-	var config godo.DatabaseLogsinkConfig
-	err := json.Unmarshal([]byte(configStr), &config)
-	if err != nil {
-		return mcp.NewToolResultError("Invalid config_json: " + err.Error()), nil
-	}
-	createReq := &godo.DatabaseCreateLogsinkRequest{
-		Name:   name,
-		Type:   typeStr,
-		Config: &config,
-	}
-	logsink, _, err := s.client.Databases.CreateLogsink(ctx, id, createReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonLogsink, err := json.MarshalIndent(logsink, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonLogsink)), nil
-}
-
-func (s *ClusterTool) getLogsink(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	logsinkID, ok := args["logsink_id"].(string)
-	if !ok || logsinkID == "" {
-		return mcp.NewToolResultError("logsink_id is required"), nil
-	}
-	logsink, _, err := s.client.Databases.GetLogsink(ctx, id, logsinkID)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonLogsink, err := json.MarshalIndent(logsink, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonLogsink)), nil
-}
-
-func (s *ClusterTool) listLogsinks(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	// Optional pagination
-	page := 0
-	if pStr, ok := args["page"].(string); ok && pStr != "" {
-		if p, err := strconv.Atoi(pStr); err == nil {
-			page = p
-		}
-	}
-	perPage := 0
-	if ppStr, ok := args["per_page"].(string); ok && ppStr != "" {
-		if pp, err := strconv.Atoi(ppStr); err == nil {
-			perPage = pp
-		}
-	}
-	var opts *godo.ListOptions
-	if page > 0 || perPage > 0 {
-		opts = &godo.ListOptions{Page: page, PerPage: perPage}
-	}
-	logsinks, _, err := s.client.Databases.ListLogsinks(ctx, id, opts)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonLogsinks, err := json.MarshalIndent(logsinks, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonLogsinks)), nil
-}
-
-func (s *ClusterTool) updateLogsink(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	logsinkID, ok := args["logsink_id"].(string)
-	if !ok || logsinkID == "" {
-		return mcp.NewToolResultError("logsink_id is required"), nil
-	}
-	configStr, ok := args["config_json"].(string)
-	if !ok || configStr == "" {
-		return mcp.NewToolResultError("config_json is required (JSON for DatabaseLogsinkConfig)"), nil
-	}
-	var config godo.DatabaseLogsinkConfig
-	err := json.Unmarshal([]byte(configStr), &config)
-	if err != nil {
-		return mcp.NewToolResultError("Invalid config_json: " + err.Error()), nil
-	}
-	updateReq := &godo.DatabaseUpdateLogsinkRequest{Config: &config}
-	_, err = s.client.Databases.UpdateLogsink(ctx, id, logsinkID, updateReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Logsink updated successfully"), nil
-}
-
-func (s *ClusterTool) deleteLogsink(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	logsinkID, ok := args["logsink_id"].(string)
-	if !ok || logsinkID == "" {
-		return mcp.NewToolResultError("logsink_id is required"), nil
-	}
-	_, err := s.client.Databases.DeleteLogsink(ctx, id, logsinkID)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Logsink deleted successfully"), nil
-}
-
-func (s *ClusterTool) startOnlineMigration(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	sourceStr, ok := args["source_json"].(string)
-	if !ok || sourceStr == "" {
-		return mcp.NewToolResultError("source_json is required (JSON for DatabaseOnlineMigrationConfig)"), nil
-	}
-	var source godo.DatabaseOnlineMigrationConfig
-	err := json.Unmarshal([]byte(sourceStr), &source)
-	if err != nil {
-		return mcp.NewToolResultError("Invalid source_json: " + err.Error()), nil
-	}
-	disableSSL := false
-	if dssl, ok := args["disable_ssl"].(string); ok && dssl != "" {
-		if b, err := strconv.ParseBool(dssl); err == nil {
-			disableSSL = b
-		}
-	}
-	var ignoreDBs []string
-	if ignoreStr, ok := args["ignore_dbs"].(string); ok && ignoreStr != "" {
-		for _, db := range strings.Split(ignoreStr, ",") {
-			db = strings.TrimSpace(db)
-			if db != "" {
-				ignoreDBs = append(ignoreDBs, db)
-			}
-		}
-	}
-	startReq := &godo.DatabaseStartOnlineMigrationRequest{
-		Source:     &source,
-		DisableSSL: disableSSL,
-		IgnoreDBs:  ignoreDBs,
-	}
-	status, _, err := s.client.Databases.StartOnlineMigration(ctx, id, startReq)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonStatus, err := json.MarshalIndent(status, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonStatus)), nil
-}
-
-func (s *ClusterTool) stopOnlineMigration(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	migrationID, ok := args["migration_id"].(string)
-	if !ok || migrationID == "" {
-		return mcp.NewToolResultError("migration_id is required"), nil
-	}
-	_, err := s.client.Databases.StopOnlineMigration(ctx, id, migrationID)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	return mcp.NewToolResultText("Online migration stopped successfully"), nil
-}
-
-func (s *ClusterTool) getOnlineMigrationStatus(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	args := req.GetArguments()
-	id, ok := args["ID"].(string)
-	if !ok || id == "" {
-		return mcp.NewToolResultError("Cluster ID is required"), nil
-	}
-	status, _, err := s.client.Databases.GetOnlineMigrationStatus(ctx, id)
-	if err != nil {
-		return mcp.NewToolResultErrorFromErr("api error", err), nil
-	}
-	jsonStatus, err := json.MarshalIndent(status, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("marshal error: %w", err)
-	}
-	return mcp.NewToolResultText(string(jsonStatus)), nil
 }
 
 func (s *ClusterTool) Tools() []server.ServerTool {
