+++ internal/dbaas/db.go
@@ -1,0 +1,108 @@
+package dbaas
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"strconv"
+
+	"github.com/digitalocean/godo"
+	"github.com/mark3labs/mcp-go/mcp"
+)
+
+func (s *ClusterTool) listDBs(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+
+	// Optional pagination
+	page := 0
+	if pStr, ok := args["page"].(string); ok && pStr != "" {
+		if p, err := strconv.Atoi(pStr); err == nil {
+			page = p
+		}
+	}
+	perPage := 0
+	if ppStr, ok := args["per_page"].(string); ok && ppStr != "" {
+		if pp, err := strconv.Atoi(ppStr); err == nil {
+			perPage = pp
+		}
+	}
+	var opts *godo.ListOptions
+	if page > 0 || perPage > 0 {
+		opts = &godo.ListOptions{Page: page, PerPage: perPage}
+	}
+
+	dbs, _, err := s.client.Databases.ListDBs(ctx, id, opts)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	jsonDBs, err := json.MarshalIndent(dbs, "", "  ")
+	if err != nil {
+		return nil, fmt.Errorf("marshal error: %w", err)
+	}
+	return mcp.NewToolResultText(string(jsonDBs)), nil
+}
+
+func (s *ClusterTool) createDB(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+	name, ok := args["name"].(string)
+	if !ok || name == "" {
+		return mcp.NewToolResultError("Database name is required"), nil
+	}
+
+	createReq := &godo.DatabaseCreateDBRequest{Name: name}
+	db, _, err := s.client.Databases.CreateDB(ctx, id, createReq)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	jsonDB, err := json.MarshalIndent(db, "", "  ")
+	if err != nil {
+		return nil, fmt.Errorf("marshal error: %w", err)
+	}
+	return mcp.NewToolResultText(string(jsonDB)), nil
+}
+
+func (s *ClusterTool) getDB(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+	name, ok := args["name"].(string)
+	if !ok || name == "" {
+		return mcp.NewToolResultError("Database name is required"), nil
+	}
+	db, _, err := s.client.Databases.GetDB(ctx, id, name)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	jsonDB, err := json.MarshalIndent(db, "", "  ")
+	if err != nil {
+		return nil, fmt.Errorf("marshal error: %w", err)
+	}
+	return mcp.NewToolResultText(string(jsonDB)), nil
+}
+
+func (s *ClusterTool) deleteDB(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+	name, ok := args["name"].(string)
+	if !ok || name == "" {
+		return mcp.NewToolResultError("Database name is required"), nil
+	}
+	_, err := s.client.Databases.DeleteDB(ctx, id, name)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	return mcp.NewToolResultText("Database deleted successfully"), nil
+}
