+++ internal/dbaas/index.go
@@ -1,0 +1,84 @@
+package dbaas
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"strconv"
+	"strings"
+
+	"github.com/digitalocean/godo"
+	"github.com/mark3labs/mcp-go/mcp"
+)
+
+func (s *ClusterTool) listIndexes(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+
+	opts := &godo.ListOptions{}
+	if pStr, ok := args["page"].(string); ok && pStr != "" {
+		if p, err := strconv.Atoi(pStr); err == nil {
+			opts.Page = p
+		}
+	}
+	if ppStr, ok := args["per_page"].(string); ok && ppStr != "" {
+		if pp, err := strconv.Atoi(ppStr); err == nil {
+			opts.PerPage = pp
+		}
+	}
+	if wpStr, ok := args["with_projects"].(string); ok && wpStr != "" {
+		if wp, err := strconv.ParseBool(wpStr); err == nil {
+			opts.WithProjects = wp
+		}
+	}
+	if odStr, ok := args["only_deployed"].(string); ok && odStr != "" {
+		if od, err := strconv.ParseBool(odStr); err == nil {
+			opts.Deployed = od
+		}
+	}
+	if poStr, ok := args["public_only"].(string); ok && poStr != "" {
+		if po, err := strconv.ParseBool(poStr); err == nil {
+			opts.PublicOnly = po
+		}
+	}
+	if ucStr, ok := args["usecases"].(string); ok && ucStr != "" {
+		ucList := []string{}
+		for _, u := range strings.Split(ucStr, ",") {
+			u = strings.TrimSpace(u)
+			if u != "" {
+				ucList = append(ucList, u)
+			}
+		}
+		opts.Usecases = ucList
+	}
+
+	indexes, _, err := s.client.Databases.ListIndexes(ctx, id, opts)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	jsonIndexes, err := json.MarshalIndent(indexes, "", "  ")
+	if err != nil {
+		return nil, fmt.Errorf("marshal error: %w", err)
+	}
+	return mcp.NewToolResultText(string(jsonIndexes)), nil
+}
+
+func (s *ClusterTool) deleteIndex(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+	name, ok := args["name"].(string)
+	if !ok || name == "" {
+		return mcp.NewToolResultError("Index name is required"), nil
+	}
+	_, err := s.client.Databases.DeleteIndex(ctx, id, name)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	return mcp.NewToolResultText("Index deleted successfully"), nil
+}
