+++ internal/dbaas/postgres.go
@@ -1,0 +1,49 @@
+package dbaas
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+
+	"github.com/digitalocean/godo"
+	"github.com/mark3labs/mcp-go/mcp"
+)
+
+func (s *ClusterTool) getPostgreSQLConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+	cfg, _, err := s.client.Databases.GetPostgreSQLConfig(ctx, id)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	jsonCfg, err := json.MarshalIndent(cfg, "", "  ")
+	if err != nil {
+		return nil, fmt.Errorf("marshal error: %w", err)
+	}
+	return mcp.NewToolResultText(string(jsonCfg)), nil
+}
+
+func (s *ClusterTool) updatePostgreSQLConfig(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+	configStr, ok := args["config_json"].(string)
+	if !ok || configStr == "" {
+		return mcp.NewToolResultError("config_json is required (JSON for PostgreSQLConfig)"), nil
+	}
+	var config godo.PostgreSQLConfig
+	err := json.Unmarshal([]byte(configStr), &config)
+	if err != nil {
+		return mcp.NewToolResultError("Invalid config_json: " + err.Error()), nil
+	}
+	_, err = s.client.Databases.UpdatePostgreSQLConfig(ctx, id, &config)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	return mcp.NewToolResultText("PostgreSQL config updated successfully"), nil
+}
