+++ internal/dbaas/user.go
@@ -1,0 +1,157 @@
+package dbaas
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"strconv"
+
+	"github.com/digitalocean/godo"
+	"github.com/mark3labs/mcp-go/mcp"
+)
+
+func (s *ClusterTool) getUser(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+	user, ok := args["user"].(string)
+	if !ok || user == "" {
+		return mcp.NewToolResultError("User name is required"), nil
+	}
+
+	dbUser, _, err := s.client.Databases.GetUser(ctx, id, user)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	jsonUser, err := json.MarshalIndent(dbUser, "", "  ")
+	if err != nil {
+		return nil, fmt.Errorf("marshal error: %w", err)
+	}
+	return mcp.NewToolResultText(string(jsonUser)), nil
+}
+
+func (s *ClusterTool) listUsers(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+
+	// Optional pagination
+	page := 0
+	if pStr, ok := args["page"].(string); ok && pStr != "" {
+		if p, err := strconv.Atoi(pStr); err == nil {
+			page = p
+		}
+	}
+	perPage := 0
+	if ppStr, ok := args["per_page"].(string); ok && ppStr != "" {
+		if pp, err := strconv.Atoi(ppStr); err == nil {
+			perPage = pp
+		}
+	}
+	var opts *godo.ListOptions
+	if page > 0 || perPage > 0 {
+		opts = &godo.ListOptions{Page: page, PerPage: perPage}
+	}
+
+	users, _, err := s.client.Databases.ListUsers(ctx, id, opts)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	jsonUsers, err := json.MarshalIndent(users, "", "  ")
+	if err != nil {
+		return nil, fmt.Errorf("marshal error: %w", err)
+	}
+	return mcp.NewToolResultText(string(jsonUsers)), nil
+}
+
+func (s *ClusterTool) createUser(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+	name, ok := args["name"].(string)
+	if !ok || name == "" {
+		return mcp.NewToolResultError("User name is required"), nil
+	}
+
+	createReq := &godo.DatabaseCreateUserRequest{Name: name}
+
+	if plugin, ok := args["mysql_auth_plugin"].(string); ok && plugin != "" {
+		createReq.MySQLSettings = &godo.DatabaseMySQLUserSettings{AuthPlugin: plugin}
+	}
+
+	if settingsStr, ok := args["settings_json"].(string); ok && settingsStr != "" {
+		var settings godo.DatabaseUserSettings
+		err := json.Unmarshal([]byte(settingsStr), &settings)
+		if err != nil {
+			return mcp.NewToolResultError("Invalid settings_json: " + err.Error()), nil
+		}
+		createReq.Settings = &settings
+	}
+
+	dbUser, _, err := s.client.Databases.CreateUser(ctx, id, createReq)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	jsonUser, err := json.MarshalIndent(dbUser, "", "  ")
+	if err != nil {
+		return nil, fmt.Errorf("marshal error: %w", err)
+	}
+	return mcp.NewToolResultText(string(jsonUser)), nil
+}
+
+func (s *ClusterTool) updateUser(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+	user, ok := args["user"].(string)
+	if !ok || user == "" {
+		return mcp.NewToolResultError("User name is required"), nil
+	}
+
+	updateReq := &godo.DatabaseUpdateUserRequest{}
+
+	if settingsStr, ok := args["settings_json"].(string); ok && settingsStr != "" {
+		var settings godo.DatabaseUserSettings
+		err := json.Unmarshal([]byte(settingsStr), &settings)
+		if err != nil {
+			return mcp.NewToolResultError("Invalid settings_json: " + err.Error()), nil
+		}
+		updateReq.Settings = &settings
+	}
+
+	dbUser, _, err := s.client.Databases.UpdateUser(ctx, id, user, updateReq)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	jsonUser, err := json.MarshalIndent(dbUser, "", "  ")
+	if err != nil {
+		return nil, fmt.Errorf("marshal error: %w", err)
+	}
+	return mcp.NewToolResultText(string(jsonUser)), nil
+}
+
+func (s *ClusterTool) deleteUser(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	args := req.GetArguments()
+	id, ok := args["ID"].(string)
+	if !ok || id == "" {
+		return mcp.NewToolResultError("Cluster ID is required"), nil
+	}
+	user, ok := args["user"].(string)
+	if !ok || user == "" {
+		return mcp.NewToolResultError("User name is required"), nil
+	}
+
+	_, err := s.client.Databases.DeleteUser(ctx, id, user)
+	if err != nil {
+		return mcp.NewToolResultErrorFromErr("api error", err), nil
+	}
+	return mcp.NewToolResultText("User deleted successfully"), nil
+}
